Makefile: fix coverage target
when rebasing the feature/site branch I encountered some weirdness
when merging the coverage target. it turned out to be caused by an
inconsistency in the flags passed to istanbul. before the
"--include-all-sources" flag an empty "-x" was given, which made the
"--include-all-sources" flag an exclude pattern instead.

I fixed that, and added vender to the list of exclude patterns, and
folded out the exlude lines over multiple lines, to make it less
cramped than it was before and less likely to introduce errors
again.