Assertion.prototype.shift: Require the next argument to be a string.
This is a breaking change as functions were previously supported à la "to satisfy".

Peter Müller brought up a problem on the gitter channel where the accidental omission of an assertion string would cause a test to pass. This fix is an attempt to avoid that particular pitfall and a bunch of equivalent ones.