Implement 'to have (ansi|text|html) diff' and support expected output-generating functions as the value.
If the function doesn't generate any output via this, assume that it did assertions based on the serialized message,
which it received as the first argument.

This also moves the magicpen type into types.js, which we probably want at some point anyway. It also fixes
the comparison of magicpen instances that have been specialized to a format.