Added '<object> not to have keys' assertion
This should be a nicer way to check if an object is "empty" rather
than the '<object> to be empty' assertion which was removed in v6.
It simply delegates to '<object> to only have keys' on an empty
array of keys.
However, this commit does not add '<object> to have keys' which
doesn't seem useful.