Add support for asserting, inspecting and diffing non-numerical array properties.
Use a 2d array in <array-like> to satisfy <array-like> again.

Only consider the subject type when determining whether to include non-numerical properties and symbols.
Fixes the test where an arguments object is used as the RHS of a `<array-like> to satisfy <array-like>`
on node.js 4 and 5 where arguments objects have some weird ArrayValues symbol.

This needs to be thought through so we know it does the right thing when the RHS is a subtype
of array-like, but I think this is the way to go.

Consolidate key-value output in a 'property' style. Get rid of the key and nonLeadingSpace styles.

Update to array-changes 1.3.0.