Remove the ability for a plugin to specify required dependencies
Since this is a breaking change, it should be landed no sooner than for
Unexpected 11.

As far as I can tell, this feature is not being used, and now
that expect.child has been implemented, it's always preferable
for plugins to bundle/depend on their dependencies and install
them in a child instance.

Previous discussion: https://github.com/unexpectedjs/unexpected/pull/377#pullrequestreview-30328564