Rework handling of the promises for unique keys as an array.
When generating to promises which check satisfying on the values
for each key, store this list in an array. Use the index into the
uniqueKeys array to grab the promises when rendering the diff.

Requisite for this is the uniqueKeys array being stable, so move
uniqueKey generation higher but then avoid the extra filter pass
that was previously done.