Rework <array-like> "to satisfy" key checking to regain performance.
Rather than apply a long series of filters on top of the entirety of
the subject getKeys() result, use the fact that within a satisfying
assertion only the value keys are important to us and whittle down
the keys we may need to filter instead.

Achieve this by altering the key presence object to store subject
keys that have not been accounted for. Apply filters to only the
remaining keys, which will end up a far fewer number, and should
any remain we fail the assertion and enter the diff path.