Pass typing information for equality through Error.equal().
Add a new findAllTypesOf() utility function that returns all the
type information required for .equal().

Arrange for this to be passed into the .equal() method. Pass it
through the overridden Error type equality method but ensure that
in the absence of specific hints the type objects are defaulted
to the current type itself which ensures backwards compatibility.

This commit also removes a number of bind() calls from definitions
attached within createWrappedExpectProto because the base methods
are defined using arrow functions thus fixing this is not required.