Address a number of issues with array-like "to satisfy".
This commit fixes key checking issues when satisfying array-like.
First, when a non-numerical property comparison is forced by it
being supplied on the RHS this should work. Secondly, in the case
of a non-numerical array-like there was an issue calculating how
many outstanding keys there were when enumerability differences
came into the picture. Fixes regression tests added in f44439d.