Rework unifiedDiff to surface all newline changes for handling.
Previously the code used the change count as authoritative on the
how much we should output. This however caused ignoring some newline
characters that we may wish to output.

Rearrange things to make change.count more of a hint when writing
out allChangedLines. This requires other areas to catch up so we
expand other routine to detect whether there were newline changes
and account for their presence.